Broke out AsyncNotificationBadge class from MainActivity.java
